+++ gc.c
@@ -1,6 +1,7 @@
 
 #include <string.h>
 #include <stdio.h>
+#include <inttypes.h>
 #include <string.h>
 #include <stdlib.h>
 #include <time.h>
@@ -19,7 +20,7 @@
 #include "scq_version.h"
 
 static void print_status_gc(const char* k, int i, const char *chr, int pos){
-    log_msg("processed %i %s through %s:%i", i, k, chr, pos);
+    log_msg("processed %" PRIu64 " %s through %s:%i", i, k, chr, pos);
 }
 
 static void usage(FILE *fp, int exit_status){
@@ -246,7 +247,7 @@ int gene_count(int argc, char *argv[]){
     }
     /* */
 
-    int n_reads = 0;
+    uint64_t n_reads = 0;
     int n_bc = 0;
     /* */
 
@@ -283,9 +284,9 @@ int gene_count(int argc, char *argv[]){
     int iter_ret;
     bam_r = bam_init1();
     while ( (iter_ret = sam_itr_next(bam, bam_itr, bam_r)) >= 0){
-        if ( (verbose) && (n_reads % (int)10e6 == 0)){
+        if ( (verbose) && (n_reads % (uint64_t)10e6 == 0)){
             const char *chr = sam_hdr_tid2name(bam_hdr, bam_r->core.tid);
-            print_status_gc("million alignments", n_reads/1e6, chr, (int)(bam_r->core.pos+1));
+            print_status_gc("million alignments", n_reads/ (uint64_t)1e6, chr, (int)(bam_r->core.pos+1));
         }
         n_reads++;
